Refactored, Added Project Manager & ImGui Renderer classes
