+++ CMakeLists.txt
@@ -2,5 +2,5 @@ cmake_minimum_required(VERSION 3.30)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 project(PossumIDE LANGUAGES CXX)
 add_subdirectory(vendor/ImGuiFileDialog)
-add_subdirectory(Ferret)
 add_subdirectory(Possum)
+add_subdirectory(Ferret)
