+++ Possum/CMakeLists.txt
@@ -1,16 +1,22 @@
 set(SRCS
     src/Core/FileManager.cpp
+    src/Core/ProjectManager.cpp
+    src/Core/ImGuiRenderer.cpp
     src/Core/ImGuiCommands.cpp
-    src/PossumLayer.cpp
-    src/PossumApp.cpp
+    src/Possum/PossumLayer.cpp
+    src/Possum/PossumApp.cpp
     src/Core/Buffer.h
+    src/Core/ImGuiRenderer.h
+    src/Core/ProjectManager.h
     src/Core/ImGuiCommands.h
     src/Core/FileManager.h
-    src/PossumLayer.h
+    src/Possum/PossumLayer.h
+    src/psmpch.h
 )
 add_executable(Possum ${SRCS})
 set_property(TARGET Possum PROPERTY CXX_STANDARD 20)
 
+target_precompile_headers(Possum PUBLIC src/psmpch.h)
 target_include_directories(Possum PRIVATE
     src
     ${CMAKE_SOURCE_DIR}/Ferret/Ferret/src
