+++ Possum/src/Core/ImGuiCommands.h
@@ -10,6 +10,6 @@ namespace Ferret::Possum
     class ImGuiCommands
     {
     public:
-        static bool InputText(const char* label, CBuffer& buf, size_t buf_size, ImGuiInputTextFlags flags = 0, ImGuiInputTextCallback callback = NULL, void* user_data = NULL);
-        static bool InputTextMultiline(const char* label, CBuffer& buf, size_t buf_size, const ImVec2& size, ImGuiInputTextFlags flags = 0, ImGuiInputTextCallback callback = NULL, void* user_data = NULL);   };
+        static bool InputText(const char* label, Buffer& buf, size_t buf_size, ImGuiInputTextFlags flags = 0, ImGuiInputTextCallback callback = NULL, void* user_data = NULL);
+        static bool InputTextMultiline(const char* label, Buffer& buf, size_t buf_size, const ImVec2& size, ImGuiInputTextFlags flags = 0, ImGuiInputTextCallback callback = NULL, void* user_data = NULL);   };
 }
