+++ Possum/src/psmpch.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <iostream>
+#include <memory>
+#include <utility>
+#include <algorithm>
+#include <functional>
+
+#include <string>
+#include <sstream>
+#include <vector>
+#include <unordered_map>
+#include <map>
+#include <filesystem>
